--- IOS/ProBaseball/ProBaseball/Domain Layer/Entities/Player.swift
@@ -1,13 +1,0 @@
-//
-//  Player.swift
-//  ProBaseball
-//
-//  Created by 조중윤 on 2021/05/08.
-//
-
-import Foundation
-
-struct Player: Codable, CustomStringConvertible, Hashable {
-    let name: String
-    let plateAppearance, numberOfPitches, hitsNumbers, accumulatedOutCount: Int
-}
