+++ IOS/ProBaseball/ProBaseball/Presentation Layer/Controller/GameListViewController.swift
@@ -28,33 +28,39 @@ class GameListViewController: UIViewController {
         gameListCollectionView.register(UINib(nibName: "GameListCollectionViewCell", bundle: nil), forCellWithReuseIdentifier: "GameListCollectionViewCell")
         gameListCollectionView.dataSource = dataSource
         gameListCollectionView.delegate = self
+        self.dataSource = configureDataSource()
         
+        configureGameListLabel()
+        bind()
+    }
+    
+    func configureGameListLabel() {
         self.gameListLabel.layer.shadowColor = UIColor(named: "retroBrown")?.cgColor
         self.gameListLabel.layer.shadowOffset = CGSize(width: 0, height: 6)
                self.gameListLabel.layer.shadowOpacity = 1.0
                self.gameListLabel.layer.shadowRadius = 0.0
         
         let str = NSMutableAttributedString(string: "Game List", attributes: [
-            NSMutableAttributedString.Key.foregroundColor : UIColor(named: "retroIvory"),
-            NSMutableAttributedString.Key.strokeColor : UIColor(named: "retroBrown"),
+            NSMutableAttributedString.Key.foregroundColor : UIColor(named: "retroIvory") ?? UIColor.red,
+            NSMutableAttributedString.Key.strokeColor : UIColor(named: "retroBrown") ?? UIColor.red,
             NSMutableAttributedString.Key.strokeWidth : -5,
             NSMutableAttributedString.Key.font : UIFont.systemFont(ofSize: 80.0)
             ])
             
         gameListLabel.attributedText = str
         gameListLabel.font = UIFont(name: "AmericanCaptain", size: 100)
-        
+        upAndDownAnimation(uiView: gameListLabel)
+    }
+    
+    func upAndDownAnimation(uiView: UIView) {
         UIView.animate(withDuration:1,
         delay: 0,
         options: [.repeat, .autoreverse],
         animations: {
-            self.gameListLabel.center.y -= CGFloat(10)
+            uiView.center.y -= CGFloat(10)
         
         },
         completion: nil)
-        
-        self.dataSource = configureDataSource()
-        bind()
     }
     
     func depend(viewModel: GameListViewModel) {
@@ -64,8 +70,7 @@ class GameListViewController: UIViewController {
     func bind() {
         viewModel.fetchGameList()
         viewModel.didUpdateGameList { [weak self] (gameList) in
-//            self?.dataSource = self?.configureDataSource()
-            self?.performUIUpdate(with: gameList)
+            self?.updateSnapshot(with: gameList)
         }
     }
     
@@ -85,12 +90,6 @@ class GameListViewController: UIViewController {
         snapshot.appendSections(GameListSection.allCases)
         snapshot.appendItems(gameList.games, toSection: GameListSection.main)
         dataSource.apply(snapshot)
-    }
-    
-    func performUIUpdate(with gameList: GameList) {
-        DispatchQueue.main.async {
-            self.updateSnapshot(with: gameList)
-        }
     }
 }
